Project Start
Decided to create a python authentication library because I thought it would be cool to create my own tailored to my needs.

Started on the API by pasting it from another project I created. It uses flask to power itself. I plan to rewrite basically everything to make it ten times better.

Once the API is done, I'll probably work on the client side library after.

This commit of pyauth is currently semi-functional.