Rewrite | Major Update
I've rewritten the entirety of the code to be more cleaner and actually work. I also made the console look a lot more snazzy by adding colorama and a lot of color codes into the project.

Planning to add HWID, Location, Library version check and more to make the authentication protocol more stronger.

This is all designed to be fully configurable and able to disable and enable features and have the ability to change database types from postgres to mongo and sqlite.

Right now the project is hung up on a postgres error which i will fix tommorow.

Then i have to start writing the client side of things which is also going to be annoying

This commit of pyauth is currently non-functional.